+++ src/PhpWord/Element/Image.php
@@ -113,7 +113,7 @@ class Image extends Element
                 \IMAGETYPE_PNG, \IMAGETYPE_BMP,
                 \IMAGETYPE_TIFF_II, \IMAGETYPE_TIFF_MM
             );
-            if (!\file_exists($source)) {
+            if (!file_exists($source)) {
                 throw new InvalidImageException;
             }
             $imgData = getimagesize($source);